put transfer table in components. changed the Route on App.js for the landing page
