+++ client/src/App.js
@@ -3,7 +3,7 @@ import "./App.css";
 import { BrowserRouter as Router, Switch, Route} from "react-router-dom";
 import Login from "./pages/Login";
 import Signup from "./pages/Signup";
-import User from "./pages/user";
+import User from "./pages/Landing";
 import useMediaQuery from '@material-ui/core/useMediaQuery';
 import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
@@ -41,7 +41,7 @@ function App() {
             <Route exact path={process.env.PUBLIC_URL + "/"} component={Login} />
             <Route exact path={process.env.PUBLIC_URL + "/login"} component={Login} />
             <Route exact path={process.env.PUBLIC_URL + "/signup"} component={Signup} />
-            <Route path={process.env.PUBLIC_URL + "/:id"} component={User} />
+            <Route path={process.env.PUBLIC_URL + "/landing"} component={User} />
           </Switch>
         </Router>
       </UserProvider>
