+++ client/src/components/header/Header.js
@@ -14,7 +14,7 @@ const useStyles = makeStyles({
   },
 });
 
-export default function ImgMediaCard() {
+export default function Header() {
   const classes = useStyles();
 
   return (
